Explode parambank into many different files
