+++ src/StudioCore/ParamEditor/ParamBankUpgrader.cs
@@ -1,0 +1,378 @@
+﻿using Andre.Formats;
+using Microsoft.Extensions.Logging;
+using SoulsFormats;
+using StudioCore.Editor;
+using StudioCore.Platform;
+using StudioCore.TextEditor;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+
+namespace StudioCore.ParamEditor;
+
+/// <summary>
+///     Utilities for dealing with global params for a game
+/// </summary>
+public partial class ParamBank : DataBank
+{
+    public enum ParamUpgradeResult
+    {
+        Success = 0,
+        RowConflictsFound = -1,
+        OldRegulationNotFound = -2,
+        OldRegulationVersionMismatch = -3,
+        OldRegulationMatchesCurrent = -4
+    }
+    private enum EditOperation
+    {
+        Add,
+        Delete,
+        Modify,
+        NameChange,
+        Match
+    }
+
+    private bool _pendingUpgrade;
+
+    private static Param UpgradeParam(Param source, Param oldVanilla, Param newVanilla, HashSet<int> rowConflicts)
+    {
+        // Presorting this would make it easier, but we're trying to preserve order as much as possible
+        // Unfortunately given that rows aren't guaranteed to be sorted and there can be duplicate IDs,
+        // we try to respect the existing order and IDs as much as possible.
+
+        // In order to assemble the final param, the param needs to know where to sort rows from given the
+        // following rules:
+        // 1. If a row with a given ID is unchanged from source to oldVanilla, we source from newVanilla
+        // 2. If a row with a given ID is deleted from source compared to oldVanilla, we don't take any row
+        // 3. If a row with a given ID is changed from source compared to oldVanilla, we source from source
+        // 4. If a row has duplicate IDs, we treat them as if the rows were deduplicated and process them
+        //    in the order they appear.
+
+        // List of rows that are in source but not oldVanilla
+        Dictionary<int, List<Param.Row>> addedRows = new(source.Rows.Count);
+
+        // List of rows in oldVanilla that aren't in source
+        Dictionary<int, List<Param.Row>> deletedRows = new(source.Rows.Count);
+
+        // List of rows that are in source and oldVanilla, but are modified
+        Dictionary<int, List<Param.Row>> modifiedRows = new(source.Rows.Count);
+
+        // List of rows that only had the name changed
+        Dictionary<int, List<Param.Row>> renamedRows = new(source.Rows.Count);
+
+        // List of ordered edit operations for each ID
+        Dictionary<int, List<EditOperation>> editOperations = new(source.Rows.Count);
+
+        // First off we go through source and everything starts as an added param
+        foreach (Param.Row row in source.Rows)
+        {
+            if (!addedRows.ContainsKey(row.ID))
+            {
+                addedRows.Add(row.ID, new List<Param.Row>());
+            }
+
+            addedRows[row.ID].Add(row);
+        }
+
+        // Next we go through oldVanilla to determine if a row is added, deleted, modified, or unmodified
+        foreach (Param.Row row in oldVanilla.Rows)
+        {
+            // First off if the row did not exist in the source, it's deleted
+            if (!addedRows.ContainsKey(row.ID))
+            {
+                if (!deletedRows.ContainsKey(row.ID))
+                {
+                    deletedRows.Add(row.ID, new List<Param.Row>());
+                }
+
+                deletedRows[row.ID].Add(row);
+                if (!editOperations.ContainsKey(row.ID))
+                {
+                    editOperations.Add(row.ID, new List<EditOperation>());
+                }
+
+                editOperations[row.ID].Add(EditOperation.Delete);
+                continue;
+            }
+
+            // Otherwise the row exists in source. Time to classify it.
+            List<Param.Row> list = addedRows[row.ID];
+
+            // First we see if we match the first target row. If so we can remove it.
+            if (row.DataEquals(list[0]))
+            {
+                Param.Row modrow = list[0];
+                list.RemoveAt(0);
+                if (list.Count == 0)
+                {
+                    addedRows.Remove(row.ID);
+                }
+
+                if (!editOperations.ContainsKey(row.ID))
+                {
+                    editOperations.Add(row.ID, new List<EditOperation>());
+                }
+
+                // See if the name was not updated
+                if ((modrow.Name == null && row.Name == null) ||
+                    (modrow.Name != null && row.Name != null && modrow.Name == row.Name))
+                {
+                    editOperations[row.ID].Add(EditOperation.Match);
+                    continue;
+                }
+
+                // Name was updated
+                editOperations[row.ID].Add(EditOperation.NameChange);
+                if (!renamedRows.ContainsKey(row.ID))
+                {
+                    renamedRows.Add(row.ID, new List<Param.Row>());
+                }
+
+                renamedRows[row.ID].Add(modrow);
+
+                continue;
+            }
+
+            // Otherwise it is modified
+            if (!modifiedRows.ContainsKey(row.ID))
+            {
+                modifiedRows.Add(row.ID, new List<Param.Row>());
+            }
+
+            modifiedRows[row.ID].Add(list[0]);
+            list.RemoveAt(0);
+            if (list.Count == 0)
+            {
+                addedRows.Remove(row.ID);
+            }
+
+            if (!editOperations.ContainsKey(row.ID))
+            {
+                editOperations.Add(row.ID, new List<EditOperation>());
+            }
+
+            editOperations[row.ID].Add(EditOperation.Modify);
+        }
+
+        // Mark all remaining rows as added
+        foreach (KeyValuePair<int, List<Param.Row>> entry in addedRows)
+        {
+            if (!editOperations.ContainsKey(entry.Key))
+            {
+                editOperations.Add(entry.Key, new List<EditOperation>());
+            }
+
+            foreach (List<EditOperation> k in editOperations.Values)
+            {
+                editOperations[entry.Key].Add(EditOperation.Add);
+            }
+        }
+
+        if (editOperations.All(kvp => kvp.Value.All(eo => eo == EditOperation.Match)))
+        {
+            return oldVanilla;
+        }
+
+        Param dest = new(newVanilla);
+
+        // Now try to build the destination from the new regulation with the edit operations in mind
+        var pendingAdds = addedRows.Keys.OrderBy(e => e).ToArray();
+        var currPendingAdd = 0;
+        var lastID = 0;
+        foreach (Param.Row row in newVanilla.Rows)
+        {
+            // See if we have any pending adds we can slot in
+            while (currPendingAdd < pendingAdds.Length &&
+                   pendingAdds[currPendingAdd] >= lastID &&
+                   pendingAdds[currPendingAdd] < row.ID)
+            {
+                if (!addedRows.ContainsKey(pendingAdds[currPendingAdd]))
+                {
+                    currPendingAdd++;
+                    continue;
+                }
+
+                foreach (Param.Row arow in addedRows[pendingAdds[currPendingAdd]])
+                {
+                    dest.AddRow(new Param.Row(arow, dest));
+                }
+
+                addedRows.Remove(pendingAdds[currPendingAdd]);
+                editOperations.Remove(pendingAdds[currPendingAdd]);
+                currPendingAdd++;
+            }
+
+            lastID = row.ID;
+
+            if (!editOperations.ContainsKey(row.ID))
+            {
+                // No edit operations for this ID, so just add it (likely a new row in the update)
+                dest.AddRow(new Param.Row(row, dest));
+                continue;
+            }
+
+            // Pop the latest operation we need to do
+            EditOperation operation = editOperations[row.ID][0];
+            editOperations[row.ID].RemoveAt(0);
+            if (editOperations[row.ID].Count == 0)
+            {
+                editOperations.Remove(row.ID);
+            }
+
+            if (operation == EditOperation.Add)
+            {
+                // Getting here means both the mod and the updated regulation added a row. Our current strategy is
+                // to overwrite the new vanilla row with the modded one and add to the conflict log to give the user
+                rowConflicts.Add(row.ID);
+                dest.AddRow(new Param.Row(addedRows[row.ID][0], dest));
+                addedRows[row.ID].RemoveAt(0);
+                if (addedRows[row.ID].Count == 0)
+                {
+                    addedRows.Remove(row.ID);
+                }
+            }
+            else if (operation == EditOperation.Match)
+            {
+                // Match means we inherit updated param
+                dest.AddRow(new Param.Row(row, dest));
+            }
+            else if (operation == EditOperation.Delete)
+            {
+                // deleted means we don't add anything
+                deletedRows[row.ID].RemoveAt(0);
+                if (deletedRows[row.ID].Count == 0)
+                {
+                    deletedRows.Remove(row.ID);
+                }
+            }
+            else if (operation == EditOperation.Modify)
+            {
+                // Modified means we use the modded regulation's param
+                dest.AddRow(new Param.Row(modifiedRows[row.ID][0], dest));
+                modifiedRows[row.ID].RemoveAt(0);
+                if (modifiedRows[row.ID].Count == 0)
+                {
+                    modifiedRows.Remove(row.ID);
+                }
+            }
+            else if (operation == EditOperation.NameChange)
+            {
+                // Inherit name
+                Param.Row newRow = new(row, dest);
+                newRow.Name = renamedRows[row.ID][0].Name;
+                dest.AddRow(newRow);
+                renamedRows[row.ID].RemoveAt(0);
+                if (renamedRows[row.ID].Count == 0)
+                {
+                    renamedRows.Remove(row.ID);
+                }
+            }
+        }
+
+        // Take care of any more pending adds
+        for (; currPendingAdd < pendingAdds.Length; currPendingAdd++)
+        {
+            // If the pending add doesn't exist in the added rows list, it was a conflicting row
+            if (!addedRows.ContainsKey(pendingAdds[currPendingAdd]))
+            {
+                continue;
+            }
+
+            foreach (Param.Row arow in addedRows[pendingAdds[currPendingAdd]])
+            {
+                dest.AddRow(new Param.Row(arow, dest));
+            }
+
+            addedRows.Remove(pendingAdds[currPendingAdd]);
+            editOperations.Remove(pendingAdds[currPendingAdd]);
+        }
+
+        return dest;
+    }
+
+    // Param upgrade. Currently for Elden Ring only.
+    public ParamUpgradeResult UpgradeRegulation(ParamBank vanillaBank, string oldVanillaParamPath,
+        Dictionary<string, HashSet<int>> conflictingParams)
+    {
+        // First we need to load the old regulation
+        if (!File.Exists(oldVanillaParamPath))
+        {
+            return ParamUpgradeResult.OldRegulationNotFound;
+        }    
+        
+        // Backup modded params
+        string modRegulationPath = $@"{Project.AssetLocator.RootDirectory}\regulation.bin";
+        File.Copy(modRegulationPath, $@"{modRegulationPath}.upgrade.bak", true);
+
+        // Load old vanilla regulation
+        BND4 oldVanillaParamBnd;
+        if (Project.Type == GameType.EldenRing)
+        {
+            oldVanillaParamBnd = SFUtil.DecryptERRegulation(oldVanillaParamPath);
+        }
+        else if (Project.Type == GameType.ArmoredCoreVI)
+        {
+            oldVanillaParamBnd = SFUtil.DecryptAC6Regulation(oldVanillaParamPath);
+        }
+        else
+        {
+            throw new NotImplementedException(
+                $"Param upgrading for game type {Project.Type} is not supported.");
+        }
+
+        Dictionary<string, Param> oldVanillaParams = new();
+        ulong version;
+        LoadParamFromBinder(oldVanillaParamBnd, ref oldVanillaParams, out version, true);
+        if (version != ParamVersion)
+        {
+            return ParamUpgradeResult.OldRegulationVersionMismatch;
+        }
+
+        Dictionary<string, Param> updatedParams = new();
+        // Now we must diff everything to try and find changed/added rows for each param
+        var anyUpgrades = false;
+        foreach (var k in vanillaBank.Params.Keys)
+        {
+            // If the param is completely new, just take it
+            if (!oldVanillaParams.ContainsKey(k) || !Params.ContainsKey(k))
+            {
+                updatedParams.Add(k, vanillaBank.Params[k]);
+                continue;
+            }
+
+            // Otherwise try to upgrade
+            HashSet<int> conflicts = new();
+            Param res = UpgradeParam(Params[k], oldVanillaParams[k], vanillaBank.Params[k], conflicts);
+            if (res != oldVanillaParams[k])
+            {
+                anyUpgrades = true;
+            }
+
+            updatedParams.Add(k, res);
+
+            if (conflicts.Count > 0)
+            {
+                conflictingParams.Add(k, conflicts);
+            }
+        }
+
+        if (!anyUpgrades)
+        {
+            return ParamUpgradeResult.OldRegulationMatchesCurrent;
+        }
+
+        var oldVersion = _paramVersion;
+
+        // Set new params
+        _params = updatedParams;
+        _paramVersion = VanillaBank.ParamVersion;
+        _pendingUpgrade = true;
+
+        // Refresh dirty cache
+        UICache.ClearCaches();
+        RefreshAllParamDiffCaches(false);
+
+        return conflictingParams.Count > 0 ? ParamUpgradeResult.RowConflictsFound : ParamUpgradeResult.Success;
+    }
+}
